Refactor to core package and add tests
