+++ videocut/__init__.py
@@ -1,0 +1,2 @@
+"""Videocut package."""
+from .core import *
